Заменя double на float

+++ include/gaussian_noise.h
@@ -9,7 +9,7 @@ public:
     void add_noise(const complex_signal *input,
                    complex_signal *output,
                    int length,
-                   double snr_db,
+                   float snr_db,
                    int bits_per_symbol);
 };
 #endif // GAUSSIAN_NOISE_H
